+++ src/core/classes/Export/ExportResolver.ts
@@ -15,6 +15,7 @@ import {
   MultiGenerateSchema,
 } from "../MultiGenerate/interfaces/multiGenerate.interface.js";
 import { MultiGenerateResolver } from "../MultiGenerate/MultiGenerateResolver.js";
+import { FileFormat, FileName, Location } from "./value-object/index.js";
 
 export class ExportResolver {
   constructor(private readonly config: FileConfig) {}
@@ -34,39 +35,40 @@ export class ExportResolver {
   }
 
   private filterGenerator(): Generator {
-    if (this.config && typeof this.config.format === "string") {
-      let gen: Generator;
-      switch (this.config.format) {
-        case "json":
-          gen = new JsonGenerator(this.config);
-          break;
-        case "javascript":
-          gen = new JavascriptGenerator(this.config);
-          break;
-        case "csv":
-          gen = new CSVGenerator(this.config);
-          break;
-        case "java":
-          gen = new JavaGenerator(this.config);
-          break;
-        case "typescript":
-          gen = new TypescriptGenerator(this.config);
-          break;
-        case "yaml":
-          gen = new YamlGenerator(this.config);
-          break;
-        case "postgresql":
-          gen = new SQLGenerator(this.config);
-          break;
-        default:
-          throw new ChacaError(
-            `Format '${String(this.config.format)}' invalid`,
-          );
-      }
+    let gen: Generator;
 
-      return gen;
-    } else {
-      throw new ChacaError(`Format '${String(this.config.format)}' invalid`);
+    const config: FileConfig = {
+      fileName: new FileName(this.config.fileName).value(),
+      format: new FileFormat(this.config.format).value(),
+      location: new Location(this.config.location).value(),
+    };
+
+    switch (this.config.format) {
+      case "json":
+        gen = new JsonGenerator(config);
+        break;
+      case "javascript":
+        gen = new JavascriptGenerator(config);
+        break;
+      case "csv":
+        gen = new CSVGenerator(config);
+        break;
+      case "java":
+        gen = new JavaGenerator(config);
+        break;
+      case "typescript":
+        gen = new TypescriptGenerator(config);
+        break;
+      case "yaml":
+        gen = new YamlGenerator(config);
+        break;
+      case "postgresql":
+        gen = new SQLGenerator(config);
+        break;
+      default:
+        throw new ChacaError(`Format '${config.format}' invalid`);
     }
+
+    return gen;
   }
 }