+++ src/generators/Generator/Generator.ts
@@ -2,7 +2,6 @@ import { FileConfig } from "../../core/interfaces/export.interface.js";
 import path from "path";
 import { MultiGenerateResolver } from "../../core/classes/MultiGenerate/MultiGenerateResolver.js";
 import { ChacaUtils } from "../../core/classes/ChacaUtils/ChacaUtils.js";
-import { FileName, Location } from "./value-object/index.js";
 
 export abstract class Generator {
   protected utils = new ChacaUtils();
@@ -14,14 +13,11 @@ export abstract class Generator {
   protected baseLocation: string;
 
   constructor(extension: string, config: FileConfig) {
-    const fileName = new FileName(config.fileName).value();
-    const location = new Location(config.location).value();
-
     this.ext = extension;
     this.config = config;
-    this.fileName = `${fileName}.${this.ext}`;
-    this.baseLocation = path.join("./", location);
-    this.route = this.generateRoute(fileName);
+    this.fileName = `${this.config.fileName}.${this.ext}`;
+    this.baseLocation = path.join("./", this.config.location);
+    this.route = this.generateRoute(this.config.fileName);
   }
 
   public abstract generateFile(data: any): Promise<string>;