+++ src/generators/Generator/value-object/index.ts
@@ -1,4 +1,0 @@
-import { FileName } from "./FileName.js";
-import { Location } from "./Location.js";
-
-export { FileName, Location };